msm: mdss: fix sw_reset register write during pipe halt
sw_reset is a common register serving multiple MDP hw modules with
unique bit for each module. So whenever a write to this register is done
to reset a particular module, first read of the register needs to be
performed followed by write with modified corresponding bit. Current
implementation writes without reading, leading to unknown behaviour. Fix
this as per correct method.

Change-Id: If009e41a13d73bb8d76651d7b65ea2f5b6335da9
Signed-off-by: Ujwal Patel <ujwalp@codeaurora.org>